requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   openai
backoff==2.2.1
    # via langfuse
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.7
    # via requests
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via openai
googleapis-common-protos==1.74.0
    # via opentelemetry-exporter-otlp-proto-http
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langfuse
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
jiter==0.13.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.1.1
    # via jsonpatch
langchain==1.2.15
    # via final-project-henry (pyproject.toml)
langchain-core==1.2.27
    # via
    #   langchain
    #   langchain-openai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==1.1.12
    # via final-project-henry (pyproject.toml)
langfuse==4.0.6
    # via final-project-henry (pyproject.toml)
langgraph==1.1.6
    # via langchain
langgraph-checkpoint==4.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.9
    # via langgraph
langgraph-sdk==0.3.13
    # via langgraph
langsmith==0.7.26
    # via langchain-core
openai==2.30.0
    # via langchain-openai
opentelemetry-api==1.40.0
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.40.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.40.0
    # via langfuse
opentelemetry-proto==1.40.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.40.0
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.61b0
    # via opentelemetry-sdk
orjson==3.11.8
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   langchain-core
    #   langfuse
    #   langsmith
protobuf==6.33.6
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.12.5
    # via
    #   final-project-henry (pyproject.toml)
    #   langchain
    #   langchain-core
    #   langfuse
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.41.5
    # via pydantic
pymupdf==1.27.2.2
    # via final-project-henry (pyproject.toml)
python-dotenv==1.2.2
    # via final-project-henry (pyproject.toml)
pyyaml==6.0.3
    # via langchain-core
regex==2026.4.4
    # via tiktoken
requests==2.33.1
    # via
    #   langsmith
    #   opentelemetry-exporter-otlp-proto-http
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via openai
tenacity==9.1.4
    # via langchain-core
tiktoken==0.12.0
    # via langchain-openai
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   langchain-core
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
wrapt==1.17.3
    # via langfuse
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
zipp==3.23.0
    # via importlib-metadata
zstandard==0.25.0
    # via langsmith
